# apps/mobile/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TabScribe API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/me": {
      "get": {
        "summary": "Me",
        "description": "Who the bearer token belongs to, and how much of this month's quota is used.",
        "operationId": "me_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "me"
        ],
        "summary": "Delete Account",
        "operationId": "delete_account_me_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MeOut": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "jobs_this_month": {
            "type": "integer",
            "title": "Jobs This Month"
          },
          "jobs_per_month": {
            "type": "integer",
            "title": "Jobs Per Month"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "jobs_this_month",
          "jobs_per_month"
        ],
        "title": "MeOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}